- Eliminados color y rvcorr, quedan en standby (siempre se pueden recuperar del svn) - Se arreglo rotate y displace, se reescribio replicate - crystal2d funciona

+++ plugins/testing/replicate.h
@@ -21,7 +21,7 @@ class ReplicateModifier: public lpmd::SystemModifier, public lpmd::Plugin
   void Apply(lpmd::Simulation & con);
 
  private:
-  unsigned long nx, ny, nz;
+  int nx, ny, nz;
 };
 
 #endif
